Unit testing for ShoppingCartService

+++ src/main/java/dk/cristi/app/webshop/client/services/ShoppingCartService.java
@@ -4,23 +4,25 @@ import dk.cristi.app.webshop.client.models.entities.ShoppingCart;
 import dk.cristi.app.webshop.client.models.entities.ShoppingCartItem;
 import dk.cristi.app.webshop.client.repositories.ShoppingCartItemRepository;
 import dk.cristi.app.webshop.client.repositories.ShoppingCartRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.*;
+import java.time.ZoneOffset;
+import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
 @Service
 public class ShoppingCartService {
 
+    private static final Logger logger = LoggerFactory.getLogger(ShoppingCartService.class);
     private final ShoppingCartRepository shoppingCartRepository;
     private final ShoppingCartItemRepository shoppingCartItemRepository;
-    @Value("${shoppingCart.maxTimeSpan}")
-    private long maxTimeSpan;
 
     @Autowired
     public ShoppingCartService(ShoppingCartRepository shoppingCartRepository, ShoppingCartItemRepository shoppingCartItemRepository) {
@@ -29,14 +31,14 @@ public class ShoppingCartService {
     }
 
     public Optional<ShoppingCart> getCart(String uid) {
-        return shoppingCartRepository.findByKey(uid);
+        return shoppingCartRepository.findByUid(uid);
     }
 
     @Transactional
     public ShoppingCart saveCart(ShoppingCart cart) {
 
         // Remove the previous saved cart
-        final Optional<ShoppingCart> shoppingCartOptional = deleteShoppingCartByKey(cart.getKey());// TODO: 15-May-18 Maybe use the object?
+        final Optional<ShoppingCart> shoppingCartOptional = deleteShoppingCartByKey(cart.getUid());
         // if present, reuse the old id
         shoppingCartOptional.ifPresent(shoppingCart -> cart.setId(shoppingCart.getId()));
         // Save the new cart
@@ -46,15 +48,28 @@ public class ShoppingCartService {
 
     public boolean keyIsInUse(String key) {
         return StreamSupport.stream(shoppingCartRepository.findAll().spliterator(), false)
-                .anyMatch(cart -> cart.getKey().equals(key));
+                .anyMatch(cart -> cart.getUid().equals(key));
     }
 
+    /**
+     * Deletes shopping carts which have been stored for the amount of minutes that is passed as a parameter.
+     * @param span the duration in minutes of how long a shopping cart is allowed to be stored
+     * @return the shopping carts that have been removed
+     * */
     @Transactional
-    public List<ShoppingCart> cleanShoppingCarts() {
+    public List<ShoppingCart> cleanShoppingCarts(long span) {
+        if (span <= 0) {
+            throw new IllegalArgumentException("Span must be a positive number");
+        }
         final List<ShoppingCart> toDelete = StreamSupport.stream(shoppingCartRepository.findAll().spliterator(), false)
-                .filter(cart -> cart.getRegisteredAt().plusMinutes(maxTimeSpan).isAfter(LocalDateTime.now())) // FIXME: 15-May-18 Timezones
+                // !isAfter(...) is used to also delete carts which result in registeredAt + span == now()
+                // this is an edge case scenario, but still can be taken into consideration
+                .filter(cart -> !cart.getRegisteredAt().plusMinutes(span).isAfter(LocalDateTime.now(ZoneOffset.UTC))) // FIXME: 15-May-18 Check timezones
                 .collect(Collectors.toList());
+
         shoppingCartRepository.deleteAll(toDelete); // TODO: 02-Aug-18 Delete or mark as deleted?
+
+        // Delete the items of the shopping carts
         final List<ShoppingCartItem> items = toDelete
                 .stream()
                 .flatMap(cart -> cart.getItems().stream())
@@ -65,7 +80,7 @@ public class ShoppingCartService {
 
     @Transactional
     public Optional<ShoppingCart> deleteShoppingCartByKey(String key) {
-        final Optional<ShoppingCart> shoppingCartOptional = shoppingCartRepository.findByKey(key);
+        final Optional<ShoppingCart> shoppingCartOptional = shoppingCartRepository.findByUid(key);
         if (!shoppingCartOptional.isPresent()) {
             return shoppingCartOptional;
         }
